requirements.txt
bcrypt==3.1.4
Flask==0.12.2
Flask-JWT-Extended==3.7.2
jsonschema==2.6.0
geojson==2.3.0
gunicorn==19.7.1
healthcheck==1.3.2
pycrypto==2.6.1
pymongo==3.7.1
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
